requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
Jinja2==3.0.1
itsdangerous==2.0.1
MarkupSafe==2.0.1
SQLAlchemy==1.4.23
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
WTForms==2.3.3
email-validator==1.1.3
Pillow==10.2.0
python-dotenv==0.19.0
